src/requirements.txt
google_search_results==2.4.2
joblib==1.4.2
matplotlib==3.9.1
numpy==2.0.0
opencv_python==4.10.0.84
pandas==2.2.2
Pillow==10.4.0
python-dotenv==1.0.1
scikit_learn==1.5.1
streamlit==1.36.0
torch==2.3.1
torchvision==0.18.1
